+++ ddm/src/exchange/runtime.rs
@@ -2,19 +2,15 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-//! This file implements the ddm router prefix exchange mechanisms. These
-//! mechanisms are responsible for announcing and withdrawing prefix sets to and
-//! from peers.
-//!
-//! This file has a set of request initiators and request handlers for
-//! announcing, withdrawing and synchronizing routes with a a given peer.
-//! Communication between peers is over HTTP(s) requests.
-//!
-//! This file only contains basic mechanisms for prefix information exchange
-//! with peers. How those mechanisms are used in the overall state machine model
-//! of a ddm router is defined in the state machine implementation in sm.rs.
-//!
+//! Runtime helpers for the ddm prefix exchange protocol: HTTP push/pull
+//! initiators and dropshot endpoint handlers, and the route programming
+//! plumbing that drains received updates into the local DB and the
+//! forwarding platform via [`crate::sys`]. illumos-only.
 
+use super::{
+    ExchangeError, PullResponse, PullResponseV2, TunnelUpdate, UnderlayUpdate,
+    Update, UpdateV2,
+};
 use crate::db::{Route, effective_route_set};
 use crate::discovery::Version;
 use crate::sm::{Config, Event, PeerEvent, SmContext};
@@ -23,8 +19,6 @@ use ddm_types::db::{RouterKind, TunnelRoute};
 use ddm_types::exchange::{PathVector, PathVectorV2};
 use dropshot::ApiDescription;
 use dropshot::ConfigDropshot;
-use dropshot::ConfigLogging;
-use dropshot::ConfigLoggingLevel;
 use dropshot::HttpError;
 use dropshot::HttpResponseOk;
 use dropshot::HttpResponseUpdatedNoContent;
@@ -37,17 +31,16 @@ use hyper::body::Bytes;
 use hyper_util::client::legacy::Client;
 use hyper_util::rt::TokioExecutor;
 use mg_common::net::{TunnelOrigin, TunnelOriginV2};
-use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
-use slog::Logger;
+use slog::{Logger, o};
 use std::collections::HashSet;
 use std::net::{Ipv6Addr, SocketAddrV6};
 use std::sync::Arc;
 use std::sync::atomic::Ordering;
 use std::time::Duration;
-use thiserror::Error;
 use tokio::sync::Mutex;
 use tokio::time::timeout;
+
+const UNIT_EXCHANGE_SERVER: &str = "exchange_server";
 
 #[derive(Clone)]
 pub struct HandlerContext {
@@ -56,300 +49,16 @@ pub struct HandlerContext {
     log: Logger,
 }
 
-/// THIS TYPE IS FOR DDM PROTOCOL VERSION 1. IT SHALL NEVER CHANGE. THIS TYPE
-/// CAN BE REMOVED WHEN DDMV1 CLIENTS AND SERVERS NO LONGER EXIST BUT ITS
-/// DEFINITION SHALL NEVER CHANGE.
-#[derive(Debug, Clone, Deserialize, Serialize, JsonSchema, Default)]
-pub struct UpdateV1 {
-    pub announce: HashSet<PathVector>,
-    pub withdraw: HashSet<PathVector>,
-}
-
-/// THIS TYPE IS FOR DDM PROTOCOL VERSION 2. IT SHALL NEVER CHANGE. THIS TYPE
-/// CAN BE REMOVED WHEN DDMV2 CLIENTS AND SERVERS NO LONGER EXIST BUT ITS
-/// DEFINITION SHALL NEVER CHANGE.
-#[derive(Debug, Clone, Deserialize, Serialize, JsonSchema, Default)]
-pub struct UpdateV2 {
-    pub underlay: Option<UnderlayUpdateV2>,
-    pub tunnel: Option<TunnelUpdateV2>,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, JsonSchema, Default)]
-pub struct Update {
-    pub underlay: Option<UnderlayUpdate>,
-    pub tunnel: Option<TunnelUpdate>,
-}
-
-impl From<UpdateV1> for Update {
-    fn from(value: UpdateV1) -> Self {
-        Update {
-            tunnel: None,
-            underlay: Some(UnderlayUpdate {
-                announce: value.announce,
-                withdraw: value.withdraw,
-            }),
-        }
-    }
-}
-
-impl From<UpdateV2> for Update {
-    fn from(value: UpdateV2) -> Self {
-        Update {
-            tunnel: value.tunnel.map(TunnelUpdate::from),
-            underlay: value.underlay.map(UnderlayUpdate::from),
-        }
-    }
-}
-
-impl From<Update> for UpdateV1 {
-    fn from(value: Update) -> Self {
-        let (announce, withdraw) = match value.underlay {
-            Some(underlay) => (underlay.announce, underlay.withdraw),
-            None => (HashSet::new(), HashSet::new()),
-        };
-        UpdateV1 { announce, withdraw }
-    }
-}
-
-impl From<Update> for UpdateV2 {
-    fn from(value: Update) -> Self {
-        UpdateV2 {
-            tunnel: value.tunnel.map(TunnelUpdateV2::from),
-            underlay: value.underlay.map(UnderlayUpdateV2::from),
-        }
-    }
-}
-
-impl From<UnderlayUpdate> for Update {
-    fn from(u: UnderlayUpdate) -> Self {
-        Update {
-            underlay: Some(u),
-            tunnel: None,
-        }
-    }
-}
-
-impl From<TunnelUpdate> for Update {
-    fn from(t: TunnelUpdate) -> Self {
-        Update {
-            underlay: None,
-            tunnel: Some(t),
-        }
-    }
-}
-
 impl Update {
+    /// Build an `Update` whose underlay/tunnel halves carry the announcements
+    /// from `pr`. Used by [`pull`] to project a pull response back into the
+    /// update event stream.
     fn announce(pr: PullResponse) -> Self {
         Self {
             underlay: pr.underlay.map(UnderlayUpdate::announce),
             tunnel: pr.tunnel.map(TunnelUpdate::announce),
         }
     }
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, JsonSchema, Default)]
-pub struct PullResponse {
-    pub underlay: Option<HashSet<PathVector>>,
-    pub tunnel: Option<HashSet<TunnelOrigin>>,
-}
-
-/// THIS TYPE IS FOR DDM PROTOCOL VERSION 2. IT SHALL NEVER CHANGE. THIS TYPE
-/// CAN BE REMOVED WHEN DDMV2 CLIENTS AND SERVERS NO LONGER EXIST BUT ITS
-/// DEFINITION SHALL NEVER CHANGE.
-#[derive(Debug, Clone, Deserialize, Serialize, JsonSchema, Default)]
-pub struct PullResponseV2 {
-    pub underlay: Option<HashSet<PathVectorV2>>,
-    pub tunnel: Option<HashSet<TunnelOriginV2>>,
-}
-
-impl From<PullResponseV2> for PullResponse {
-    fn from(value: PullResponseV2) -> Self {
-        PullResponse {
-            underlay: value
-                .underlay
-                .map(|x| x.into_iter().map(PathVector::from).collect()),
-            tunnel: value
-                .tunnel
-                .map(|x| x.into_iter().map(TunnelOrigin::from).collect()),
-        }
-    }
-}
-
-impl From<HashSet<PathVector>> for PullResponse {
-    fn from(value: HashSet<PathVector>) -> Self {
-        PullResponse {
-            underlay: Some(value),
-            tunnel: None,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, JsonSchema, Default)]
-pub struct UnderlayUpdate {
-    pub announce: HashSet<PathVector>,
-    pub withdraw: HashSet<PathVector>,
-}
-
-/// THIS TYPE IS FOR DDM PROTOCOL VERSION 2. IT SHALL NEVER CHANGE. THIS TYPE
-/// CAN BE REMOVED WHEN DDMV2 CLIENTS AND SERVERS NO LONGER EXIST BUT ITS
-/// DEFINITION SHALL NEVER CHANGE.
-#[derive(Debug, Clone, Deserialize, Serialize, JsonSchema, Default)]
-pub struct UnderlayUpdateV2 {
-    pub announce: HashSet<PathVectorV2>,
-    pub withdraw: HashSet<PathVectorV2>,
-}
-
-impl From<UnderlayUpdate> for UnderlayUpdateV2 {
-    fn from(value: UnderlayUpdate) -> Self {
-        UnderlayUpdateV2 {
-            announce: value
-                .announce
-                .into_iter()
-                .map(PathVectorV2::from)
-                .collect(),
-            withdraw: value
-                .withdraw
-                .into_iter()
-                .map(PathVectorV2::from)
-                .collect(),
-        }
-    }
-}
-
-impl From<UnderlayUpdateV2> for UnderlayUpdate {
-    fn from(value: UnderlayUpdateV2) -> Self {
-        UnderlayUpdate {
-            announce: value
-                .announce
-                .into_iter()
-                .map(PathVector::from)
-                .collect(),
-            withdraw: value
-                .withdraw
-                .into_iter()
-                .map(PathVector::from)
-                .collect(),
-        }
-    }
-}
-
-impl UnderlayUpdate {
-    pub fn announce(prefixes: HashSet<PathVector>) -> Self {
-        Self {
-            announce: prefixes,
-            ..Default::default()
-        }
-    }
-    pub fn withdraw(prefixes: HashSet<PathVector>) -> Self {
-        Self {
-            withdraw: prefixes,
-            ..Default::default()
-        }
-    }
-    pub fn with_path_element(&self, element: String) -> Self {
-        Self {
-            announce: self
-                .announce
-                .iter()
-                .map(|x| {
-                    let mut pv = x.clone();
-                    pv.path.push(element.clone());
-                    pv
-                })
-                .collect(),
-            withdraw: self
-                .withdraw
-                .iter()
-                .map(|x| {
-                    let mut pv = x.clone();
-                    pv.path.push(element.clone());
-                    pv
-                })
-                .collect(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, JsonSchema, Default)]
-pub struct TunnelUpdate {
-    pub announce: HashSet<TunnelOrigin>,
-    pub withdraw: HashSet<TunnelOrigin>,
-}
-
-/// THIS TYPE IS FOR DDM PROTOCOL VERSION 2. IT SHALL NEVER CHANGE. THIS TYPE
-/// CAN BE REMOVED WHEN DDMV2 CLIENTS AND SERVERS NO LONGER EXIST BUT ITS
-/// DEFINITION SHALL NEVER CHANGE.
-#[derive(Debug, Clone, Deserialize, Serialize, JsonSchema, Default)]
-pub struct TunnelUpdateV2 {
-    pub announce: HashSet<TunnelOriginV2>,
-    pub withdraw: HashSet<TunnelOriginV2>,
-}
-
-impl From<TunnelUpdateV2> for TunnelUpdate {
-    fn from(value: TunnelUpdateV2) -> Self {
-        TunnelUpdate {
-            announce: value
-                .announce
-                .into_iter()
-                .map(TunnelOrigin::from)
-                .collect(),
-            withdraw: value
-                .withdraw
-                .into_iter()
-                .map(TunnelOrigin::from)
-                .collect(),
-        }
-    }
-}
-
-impl From<TunnelUpdate> for TunnelUpdateV2 {
-    fn from(value: TunnelUpdate) -> Self {
-        TunnelUpdateV2 {
-            announce: value
-                .announce
-                .into_iter()
-                .map(TunnelOriginV2::from)
-                .collect(),
-            withdraw: value
-                .withdraw
-                .into_iter()
-                .map(TunnelOriginV2::from)
-                .collect(),
-        }
-    }
-}
-
-impl TunnelUpdate {
-    pub fn announce(prefixes: HashSet<TunnelOrigin>) -> Self {
-        Self {
-            announce: prefixes,
-            ..Default::default()
-        }
-    }
-    pub fn withdraw(prefixes: HashSet<TunnelOrigin>) -> Self {
-        Self {
-            withdraw: prefixes,
-            ..Default::default()
-        }
-    }
-}
-
-#[derive(Error, Debug)]
-pub enum ExchangeError {
-    #[error("io error: {0}")]
-    Io(#[from] std::io::Error),
-
-    #[error("hyper error: {0}")]
-    Hyper(#[from] hyper::Error),
-
-    #[error("hyper client error: {0}")]
-    HyperClient(#[from] hyper_util::client::legacy::Error),
-
-    #[error("timeout error: {0}")]
-    Timeout(#[from] tokio::time::error::Elapsed),
-
-    #[error("json error: {0}")]
-    SerdeJson(#[from] serde_json::Error),
 }
 
 pub(crate) fn announce_underlay(
@@ -587,11 +296,11 @@ pub fn handler(
         ..Default::default()
     };
 
-    let ds_log = ConfigLogging::StderrTerminal {
-        level: ConfigLoggingLevel::Error,
-    }
-    .to_logger("exchange")
-    .map_err(|e| e.to_string())?;
+    let ds_log = log.new(o!(
+        "component" => crate::COMPONENT_DDM,
+        "module" => crate::MOD_EXCHANGE,
+        "unit" => UNIT_EXCHANGE_SERVER,
+    ));
 
     inf!(log, ctx.config.if_name, "exchange: listening on {}", sa);
