+++ ddm/src/lib.rs
@@ -8,8 +8,12 @@ pub mod discovery;
 pub mod exchange;
 pub mod oxstats;
 pub mod sm;
+#[cfg(all(feature = "illumos", target_os = "illumos"))]
 pub mod sys;
-mod util;
+
+pub const COMPONENT_DDM: &str = "ddm";
+pub const MOD_ADMIN: &str = "admin";
+pub const MOD_EXCHANGE: &str = "exchange";
 
 #[macro_export]
 macro_rules! err {
