+++ ddm/src/sm/state.rs
@@ -2,194 +2,32 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-use crate::db::Db;
-use crate::discovery::Version;
+//! Routing state machine implementation. The `Init` -> `Solicit` ->
+//! `Exchange` lifecycle drives kernel route programming via [`crate::sys`]
+//! and reads interface addressing through `libnet`. This module is
+//! illumos-only.
+
+use super::{
+    AdminEvent, Event, NeighborEvent, PeerEvent, PrefixSet, SmContext, SmError,
+    StateMachine,
+};
 use crate::exchange::{TunnelUpdate, UnderlayUpdate, Update};
 use crate::{dbg, discovery, err, exchange, inf, wrn};
 use ddm_types::db::RouterKind;
 use ddm_types::exchange::PathVector;
 use libnet::get_ipaddr_info;
-use mg_common::net::TunnelOrigin;
-use oxnet::Ipv6Net;
 use slog::Logger;
 use std::collections::HashSet;
-use std::net::{IpAddr, Ipv6Addr};
-use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
-use std::sync::mpsc::{Receiver, Sender};
-use std::sync::{Arc, Mutex};
-use std::thread::sleep;
-use std::thread::spawn;
+use std::net::IpAddr;
+use std::sync::Arc;
+use std::sync::atomic::{AtomicBool, Ordering};
+use std::sync::mpsc::Receiver;
+use std::thread::{sleep, spawn};
 use std::time::Duration;
-use thiserror::Error;
 
-#[derive(Debug)]
-pub enum AdminEvent {
-    /// Announce a set of IPv6 prefixes
-    Announce(PrefixSet),
-
-    /// Withdraw a set of IPv6 prefixes
-    Withdraw(PrefixSet),
-
-    /// Expire the peer at the specified address
-    Expire(Ipv6Addr),
-
-    /// Synchronize with active peers by pulling their prefixes.
-    Sync,
-}
-
-#[derive(Debug)]
-pub enum PrefixSet {
-    Underlay(HashSet<Ipv6Net>),
-    Tunnel(HashSet<TunnelOrigin>),
-}
-
-#[derive(Debug)]
-pub enum PeerEvent {
-    Push(Update),
-}
-
-#[derive(Debug)]
-pub enum NeighborEvent {
-    Advertise((Ipv6Addr, Version)),
-    SolicitFail,
-    Expire,
-}
-
-#[derive(Debug)]
-pub enum Event {
-    Neighbor(NeighborEvent),
-    Peer(PeerEvent),
-    Admin(AdminEvent),
-}
-
-impl From<NeighborEvent> for Event {
-    fn from(e: NeighborEvent) -> Self {
-        Self::Neighbor(e)
-    }
-}
-
-impl From<PeerEvent> for Event {
-    fn from(e: PeerEvent) -> Self {
-        Self::Peer(e)
-    }
-}
-
-impl From<AdminEvent> for Event {
-    fn from(e: AdminEvent) -> Self {
-        Self::Admin(e)
-    }
-}
-
-#[derive(Debug)]
-pub enum StateType {
-    Solicit,
-    Exchange,
-}
-
-#[derive(Debug)]
-pub enum EventError {
-    InvalidEvent(StateType),
-}
-
-#[derive(Debug)]
-pub enum EventResponse {
-    Success,
-    Prefixes(Vec<Ipv6Net>),
-}
-
-#[derive(Error, Debug)]
-pub enum SmError {
-    #[error("io error: {0}")]
-    Io(#[from] std::io::Error),
-
-    #[error("discovery error: {0}")]
-    Discovery(#[from] discovery::DiscoveryError),
-}
-
-#[derive(Clone)]
-pub struct Config {
-    /// Interface this state machine is associated with.
-    pub if_index: u32,
-
-    /// Interface name this state machine is associated with.
-    pub if_name: String,
-
-    /// Address object name the state machine uses for peering. Must correspond
-    /// to IPv6 link local address.
-    pub aobj_name: String,
-
-    /// Link local Ipv6 address this state machine is associated with
-    pub addr: Ipv6Addr,
-
-    /// How long to wait between solicitations (milliseconds).
-    pub solicit_interval: u64,
-
-    /// How often to check for link failure while waiting for discovery messges.
-    pub discovery_read_timeout: u64,
-
-    /// How long to wait between attempts to get an IP address for a specified
-    /// address object.
-    pub ip_addr_wait: u64,
-
-    /// How long to wait without a solicitation response before expiring a peer
-    /// (milliseconds).
-    pub expire_threshold: u64,
-
-    /// How long to wait for a response to exchange messages.
-    pub exchange_timeout: u64,
-
-    /// The kind of router this is, server or transit.
-    pub kind: RouterKind,
-
-    /// TCP port to use for prefix exchange.
-    pub exchange_port: u16,
-
-    /// Dendrite dpd config
-    pub dpd: Option<DpdConfig>,
-}
-
-#[derive(Clone)]
-pub struct DpdConfig {
-    pub host: String,
-    pub port: u16,
-}
-
-#[derive(Default)]
-pub struct SessionStats {
-    // Discovery
-    pub solicitations_sent: AtomicU64,
-    pub solicitations_received: AtomicU64,
-    pub advertisements_sent: AtomicU64,
-    pub advertisements_received: AtomicU64,
-    pub peer_expirations: AtomicU64,
-    pub peer_address_changes: AtomicU64,
-    pub peer_established: AtomicU64,
-    pub peer_address: Mutex<Option<Ipv6Addr>>,
-
-    // Exchange
-    pub updates_sent: AtomicU64,
-    pub updates_received: AtomicU64,
-    pub imported_underlay_prefixes: AtomicU64,
-    pub imported_tunnel_endpoints: AtomicU64,
-    pub update_send_fail: AtomicU64,
-}
-
-#[derive(Clone)]
-pub struct SmContext {
-    pub config: Config,
-    pub db: Db,
-    pub tx: Sender<Event>,
-    pub event_channels: Vec<Sender<Event>>,
-    pub rt: Arc<tokio::runtime::Handle>,
-    pub hostname: String,
-    pub stats: Arc<SessionStats>,
-    pub log: Logger,
-}
-
-pub struct StateMachine {
-    pub ctx: SmContext,
-    pub rx: Option<Receiver<Event>>,
-}
+use crate::discovery::Version;
+use mg_common::net::TunnelOrigin;
+use std::net::Ipv6Addr;
 
 impl StateMachine {
     pub fn run(&mut self) -> Result<(), SmError> {
@@ -369,8 +207,8 @@ impl State for Solicit {
     }
 }
 
-pub struct Exchange {
-    pub peer: Ipv6Addr,
+struct Exchange {
+    peer: Ipv6Addr,
     version: Version,
     ctx: SmContext,
     log: Logger,
