+++ ddmd/src/main.rs
@@ -6,11 +6,13 @@ use clap::Parser;
 use ddm::admin::{HandlerContext, RouterStats};
 use ddm::db::Db;
 use ddm::sm::{DpdConfig, SmContext, StateMachine};
+#[cfg(all(feature = "illumos", target_os = "illumos"))]
 use ddm::sys::Route;
 use ddm_types::db::RouterKind;
 use signal::handle_signals;
-use slog::{Drain, Logger, error};
+use slog::{Drain, Logger, error, warn};
 use std::net::{IpAddr, Ipv6Addr};
+#[cfg(all(feature = "illumos", target_os = "illumos"))]
 use std::sync::mpsc::channel;
 use std::sync::{Arc, Mutex};
 use uuid::Uuid;
@@ -100,6 +102,15 @@ struct Arg {
     /// Id of the sled this router is running on.
     #[arg(long)]
     sled_uuid: Option<Uuid>,
+
+    /// Skip the routing state machine (discovery, exchange, route
+    /// synchronization). Only the admin API server runs, allowing test
+    /// fixtures to obtain a real ddmd admin endpoint without the kernel-level
+    /// networking the state machine requires.
+    ///
+    /// Analogous to `mgd --no-bgp-dispatcher`.
+    #[arg(long, default_value_t = false)]
+    no_state_machine: bool,
 }
 
 #[derive(Debug, Parser, Clone)]
@@ -121,10 +132,7 @@ async fn run() {
         .await
         .expect("set up refresh signal handler");
 
-    let mut event_channels = Vec::new();
     let db = Db::new(&format!("{}/ddmdb", arg.data_dir), log.clone()).unwrap();
-
-    let mut sms = Vec::new();
 
     let dpd = match arg.dendrite {
         true => Some(DpdConfig {
@@ -140,54 +148,10 @@ async fn run() {
         .to_string_lossy()
         .to_string();
 
-    for name in arg.addresses {
-        let (tx, rx) = channel();
-        let config = ddm::sm::Config {
-            solicit_interval: arg.solicit_interval,
-            expire_threshold: arg.expire_threshold,
-            discovery_read_timeout: arg.discovery_read_timeout,
-            ip_addr_wait: arg.ip_addr_wait,
-            exchange_timeout: arg.exchange_timeout,
-            exchange_port: arg.exchange_port,
-            aobj_name: name.clone(),
-            if_name: String::new(),
-            if_index: 0,
-            kind: arg.kind,
-            dpd: dpd.clone(),
-            addr: Ipv6Addr::UNSPECIFIED,
-        };
-        let ctx = SmContext {
-            config,
-            db: db.clone(),
-            event_channels: Vec::new(),
-            tx: tx.clone(),
-            log: log.clone(),
-            hostname: hostname.clone(),
-            rt: rt.clone(),
-            stats: Arc::new(ddm::sm::SessionStats::default()),
-        };
-        let sm = StateMachine { ctx, rx: Some(rx) };
-        sms.push(sm);
-        event_channels.push(tx);
-    }
+    let (sms, event_channels) =
+        start_state_machines(&arg, &db, &dpd, &hostname, &rt, &log);
 
-    // Add an event channel sender for each state machine to every other state
-    // machine.
-    for (i, sm) in sms.iter_mut().enumerate() {
-        for (j, e) in event_channels.iter().enumerate() {
-            // dont give a state machine an event sender to itself.
-            if i == j {
-                continue;
-            }
-            sm.ctx.event_channels.push(e.clone());
-        }
-    }
-
-    for sm in &mut sms {
-        sm.run().unwrap();
-    }
-
-    termination_handler(db.clone(), dpd, rt, log.clone());
+    termination_handler(db.clone(), dpd.clone(), rt.clone(), log.clone());
 
     let router_stats = Arc::new(RouterStats::default());
     let peers: Vec<SmContext> = sms.iter().map(|x| x.ctx.clone()).collect();
@@ -237,6 +201,128 @@ async fn run() {
     std::thread::park();
 }
 
+/// Build, wire, and start the per-address routing state machines.
+///
+/// Returns the running [`StateMachine`] handles plus the sender side of each
+/// machine's event channel. When `--no-state-machine` is set the function
+/// short-circuits to empty vectors, leaving the daemon to serve only its
+/// admin API. The illumos and non-illumos variants share that early-exit
+/// branch; only the actual machine setup is platform-specific.
+#[cfg(all(feature = "illumos", target_os = "illumos"))]
+fn start_state_machines(
+    arg: &Arg,
+    db: &Db,
+    dpd: &Option<DpdConfig>,
+    hostname: &str,
+    rt: &Arc<tokio::runtime::Handle>,
+    log: &Logger,
+) -> (
+    Vec<StateMachine>,
+    Vec<std::sync::mpsc::Sender<ddm::sm::Event>>,
+) {
+    if arg.no_state_machine {
+        if !arg.addresses.is_empty() {
+            warn!(
+                log,
+                "--no-state-machine set; ignoring {} --addr value(s)",
+                arg.addresses.len(),
+            );
+        }
+        return (Vec::new(), Vec::new());
+    }
+
+    let mut sms = Vec::new();
+    let mut event_channels = Vec::new();
+
+    for name in &arg.addresses {
+        let (tx, rx) = channel();
+
+        let config = ddm::sm::Config {
+            solicit_interval: arg.solicit_interval,
+            expire_threshold: arg.expire_threshold,
+            discovery_read_timeout: arg.discovery_read_timeout,
+            ip_addr_wait: arg.ip_addr_wait,
+            exchange_timeout: arg.exchange_timeout,
+            exchange_port: arg.exchange_port,
+            aobj_name: name.clone(),
+            if_name: String::new(),
+            if_index: 0,
+            kind: arg.kind,
+            dpd: dpd.clone(),
+            addr: Ipv6Addr::UNSPECIFIED,
+        };
+
+        let ctx = SmContext {
+            config,
+            db: db.clone(),
+            event_channels: Vec::new(),
+            tx: tx.clone(),
+            log: log.clone(),
+            hostname: hostname.to_string(),
+            rt: rt.clone(),
+            stats: Arc::new(ddm::sm::SessionStats::default()),
+        };
+
+        let sm = StateMachine { ctx, rx: Some(rx) };
+        sms.push(sm);
+        event_channels.push(tx);
+    }
+
+    // Add an event channel sender for each state machine to every other state
+    // machine.
+    for (i, sm) in sms.iter_mut().enumerate() {
+        for (j, e) in event_channels.iter().enumerate() {
+            // dont give a state machine an event sender to itself.
+            if i == j {
+                continue;
+            }
+            sm.ctx.event_channels.push(e.clone());
+        }
+    }
+
+    for sm in &mut sms {
+        sm.run().unwrap();
+    }
+
+    (sms, event_channels)
+}
+
+/// Non-illumos variant: the routing state machine depends on illumos
+/// kernel networking, so on every other platform the function logs a warning
+/// and returns empty vectors. Test fixtures should pass `--no-state-machine`
+/// to silence the warning.
+#[cfg(not(all(feature = "illumos", target_os = "illumos")))]
+fn start_state_machines(
+    arg: &Arg,
+    _db: &Db,
+    _dpd: &Option<DpdConfig>,
+    _hostname: &str,
+    _rt: &Arc<tokio::runtime::Handle>,
+    log: &Logger,
+) -> (
+    Vec<StateMachine>,
+    Vec<std::sync::mpsc::Sender<ddm::sm::Event>>,
+) {
+    if !arg.no_state_machine {
+        warn!(
+            log,
+            "routing state machine is not available on non-illumos builds; \
+             behaving as if `--no-state-machine` were set",
+        );
+    }
+    if !arg.addresses.is_empty() {
+        warn!(
+            log,
+            "--no-state-machine set; ignoring {} --addr value(s)",
+            arg.addresses.len(),
+        );
+    }
+    (Vec::new(), Vec::new())
+}
+
+/// Install a Ctrl-C handler that withdraws ddmd's imported routes from the
+/// kernel before exiting. illumos-only.
+#[cfg(all(feature = "illumos", target_os = "illumos"))]
 fn termination_handler(
     db: Db,
     dendrite: Option<DpdConfig>,
@@ -267,6 +353,24 @@ fn termination_handler(
             error!(log, "shutdown tunnel routes: {e}");
         }
 
+        std::process::exit(SIGTERM_EXIT);
+    });
+}
+
+/// Non-illumos variant: there are no kernel routes to withdraw on these
+/// platforms, so the handler installs a Ctrl-C task that just exits cleanly.
+#[cfg(not(all(feature = "illumos", target_os = "illumos")))]
+fn termination_handler(
+    _db: Db,
+    _dendrite: Option<DpdConfig>,
+    _rt: Arc<tokio::runtime::Handle>,
+    _log: Logger,
+) {
+    tokio::spawn(async {
+        tokio::signal::ctrl_c()
+            .await
+            .expect("error setting termination handler");
+        const SIGTERM_EXIT: i32 = 130;
         std::process::exit(SIGTERM_EXIT);
     });
 }
